Add files via upload
code_uptoPCA and clustering_annotation_0504 are the two main python scripts. The original input file is too big and cannot be uploaded on github. tmp, tmp2 and tmp3 are csv files that show the cell_type composition of leiden clusters at different stages. WBG2.gmt and 160MarkerGeneList are the short and long marker gene-by-celltype tables that scoreCT uses to annotate the clusters.